LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/generated/worker-dedicated-importscripts.sub-expected.txt

CONSOLE MESSAGE: NetworkError: Load failed
CONSOLE MESSAGE: NetworkError: Load failed
CONSOLE MESSAGE: NetworkError: Load failed

Harness Error (FAIL), message = NetworkError: Load failed

PASS sec-fetch-site - Not sent to non-trustworthy same-origin destination
PASS sec-fetch-site - Not sent to non-trustworthy same-site destination
PASS sec-fetch-site - Not sent to non-trustworthy cross-site destination
PASS sec-fetch-mode - Not sent to non-trustworthy same-origin destination
PASS sec-fetch-mode - Not sent to non-trustworthy same-site destination
PASS sec-fetch-mode - Not sent to non-trustworthy cross-site destination
PASS sec-fetch-dest - Not sent to non-trustworthy same-origin destination
PASS sec-fetch-dest - Not sent to non-trustworthy same-site destination
PASS sec-fetch-dest - Not sent to non-trustworthy cross-site destination
PASS sec-fetch-user - Not sent to non-trustworthy same-origin destination
PASS sec-fetch-user - Not sent to non-trustworthy same-site destination
PASS sec-fetch-user - Not sent to non-trustworthy cross-site destination
FAIL sec-fetch-site - HTTPS downgrade (header not sent) promise_test: Unhandled rejection with value: object "[object ErrorEvent]"
FAIL sec-fetch-site - HTTPS upgrade promise_test: Unhandled rejection with value: object "[object ErrorEvent]"
FAIL sec-fetch-site - HTTPS downgrade-upgrade promise_test: Unhandled rejection with value: object "[object ErrorEvent]"
